requirements.txt
# Auto-generated from config/versions.yaml
# Do not edit manually — run: python scripts/generate_requirements.py

fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
sqlalchemy==2.0.36
alembic==1.14.0
rich==13.9.4
pydantic==2.10.3
pydantic-settings==2.7.0
pyyaml==6.0.2
psycopg2-binary==2.9.9
